Allow disabling SemVerTagNameFilter
`SemVerTagNameFilter.Apply()` doesn't check if the filter was enabled
and gets applied always. This change adds a check which allows disabling
SemVerTagNameFilter, making the default configuration of gcrgc to delete
semver tags as well.
Based on the similar check in `UntaggedFilter`.
Updated an existing test that uses a disabled `UntaggedFilter` to also
have a disabled `SemVerTagNameFilter` to verify the fix.

Signed-off-by: Sunny <darkowlzz@protonmail.com>